% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{calc_stats}
\alias{calc_stats}
\title{Title}
\usage{
calc_stats(df, x)
}
\arguments{
\item{df}{a dataframe}

\item{x}{a string indicating the statistic to calculate. Choose either "mean",
"median", or "sd"}
}
\value{
A tibble of the dataframe with \code{x}, the statistic of choice
}
\description{
This function calculates a passed in statistic (mean, median, or standard deviation)
for average medicare payments of DRG data.
}
\examples{

drg_data = read.csv("drg.csv")

calc_stats(drg_data, "mean")
}
