% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_drg_box.R
\name{make_drg_box}
\alias{make_drg_box}
\title{Title}
\usage{
make_drg_box(df, variable)
}
\arguments{
\item{df}{a dataframe}

\item{variable}{a string for the y variable. Please choose between:
"Average.Medicare.Payments", "Average.Total.PAyments", and "Average.Covered.Charges"}
}
\value{
A plot with \code{variable} versus DRG code
}
\description{
This function produces a ggplot boxplot of \code{variable} versus DRG Code
}
\examples{

drg_data = read.csv("drg.csv")

make_drg_box(drg_data, "Average.Medicare.Payments")

}
